+++ CHANGELOG.md
@@ -18,6 +18,23 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/).
 
 ### Security
 
+## v10.16.0 - 2021-03-11
+
+### Added
+
+- グローバル設定に注文変更時通知設定を追加
+- 注文にbrokerを追加
+
+### Changed
+
+- update @chevre/api-nodejs-client
+- update @cinerino/factory
+- プロジェクトメンバーの権限検索時にロールリストも返すように調整
+- 注文取引でcustomerを指定できるように調整
+- Chevre返金時にpurposeを返品アクションに指定するように調整
+- Chevreエラーハンドリング強化
+- 返金後メール送信を指定されなければ実行しないように調整
+
 ## v10.15.1 - 2021-02-11
 
 ### Changed